Added EGL_RENDERABLE_TYPE in EGL config selection
One could use EGL to create a big GL rendering context
or OpenVG rendering context. So the configuration
selection process should then choose a config that supports
the use of these other apis.
However the current config attributes do not have this parameter
so we're making assumption of always wanting to use GLES 1/2/3