WGL refactoring and fixes
WGL has been refactored and corrected. The implementation had several
errors such as using pixel formats incorrectly and assuming that the
pixel format integer values are portable across DCs. This was incorrect.

Instead on Windows and WGL the visualid is really  PIXELFORMATDESCRIPTOR
object which can be used to carry the parameters and then resolved to an
integer ("visualid" so to speak), but this is specific that DC only!

Also the functions returned by wglGetProcAddress are specific to a context.

In addition there's a need for a workaround to check if the function is part
of the original OpenGL 1x API implemented by Microsoft since some drivers
(such as Intel) do not return these functions through wglGetProcAddress.

This commit fixes issues #2 and #3