Resolve issues regarding visualid based config selection
Since we don't have a portable mechanism across all windowing
API implementations to support visualid based config selection
it's best to refactor the implementations and API to remove this
broken feature.

GLX and WGL implementations could both be made work through an
abstraction over what type is "visualid" but this would not work
on EGL implementation portably. It would work on Linux but not on
Windows since EGL uses an EGLInt which isn't meaningful on Windows.
Presumably this is the reason why for example ImgTech's GLES
implementation simply gives 0 for the visual id.

In either case the most straigtforward way to create a context is
to simply choose the desired config attributes, create a config
and then use that config to create context and later the surface.
This will always work on every platform.

This fixes issues #6 and #4