Add a theoretical support for rendering to a bitmap on Win32
Add a theoretical support for rendering to a bitmap on Win32. This
implementation is not yet working. When using the bitmap based
surface in a call to MakeCurrent the call will fail (and throw an
exception).

A working implementation would mostly be for the sake of completeness
and for having feature parity with Linux.

To complete the feature some more research should be done to figure
out the appropriate settings for creating the HDCs and contexts. It
is possible that this feature is in fact no longer supported on modern
rendering systems.