Use fibers to pump the PeekEvent message loop on Win32
Use fibers to pump the PeekEvent message loop on Win32.

The fundamental problem is that when a window is being resized/moved
the DefWindowProc enters a so called "modal window loop" and does
not return until the operation is finished. This will block the
calling thread and this is bad news for an interactive game type
of application that renders continuously and doesn't care about
WM_PAINT handling.

The current (maybe a bit experimental) solution is to conver the
calling thread into a fiber and offload the PeekEvent loop to a
separate fiber. We then set a timer when the modal loop is started
and on each timer callback yield to the main fiber.

See more detailed comment in win32/system.cpp PeekEvent function.