Add functionality to show/hide mouse cursor
Add functionality to show/hide mouse cursor. The implementation is
a kludge on both Win32 and on X11 but for different reassons. On
X11 there's no API to hide the cursor so we must create a transparent
pixmap that is to be used as the cursor. This is likely broken in some
interesting way when some additional screens, another  WM or toolkit
is thrown into the mix.

On Win32 the cursor is not per window but per application. The
implementation tries to emulate per window cursor state functionality
and is also likely broken if anything messes with the global cursor
state.