Fix incorrect initial key modifier value in TranslateKeydownEvent
Fix incorrect initial key modifier value in TranslateKeydownEvent.
The problem is that the code uses Keymod::None as the initial value
but the object created is actually a bitflag<Keymod> object which
means that the value None is stored in the 1st bit in the bitflag.
This means that the "intuitive" code for checking for checking for
"no modifier bits set" i.e. comparińg against 0 bitflag value is
actually incorrect. This fix applies to both X11 and Win32.